+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 - `markdownPreferences` via [`eslint-plugin-markdown-preferences`](https://npmjs.com/eslint-plugin-markdown-preferences), enabled by default.
 - `markdownLinks` via [`eslint-plugin-markdown-links`](https://npmjs.com/eslint-plugin-markdown-links), enabled by default.
 - `zod` via [`eslint-plugin-zod-x`](https://npmjs.com/eslint-plugin-zod-x), enabled automatically if `zod@>=4` package is installed.
+- `fast-import` via [`eslint-plugin-fast-import`](https://npmjs.com/eslint-plugin-fast-import), **<u>disabled</u>** by default.
+- `checkFile` via [`eslint-plugin-check-file`](https://npmjs.com/eslint-plugin-check-file), **<u>disabled</u>** by default.
 
 ### Changes